+++ src/CommandLineUtils/Attributes/OptionAttributeBase.cs
@@ -62,10 +62,12 @@ namespace McMaster.Extensions.CommandLineUtils
             option.ValueName = ValueName ?? option.ValueName;
             option.SymbolName = SymbolName ?? option.SymbolName;
 
+#pragma warning disable 618
             if (option.Template == null)
             {
                 option.Template = option.ToTemplateString();
             }
+#pragma warning restore 618
         }
     }
 }